+++ test/runtests.jl
@@ -42,6 +42,7 @@ const RUN_E2E = get(ENV, "MCP_TEST_E2E", ON_CI ? "false" : "true") == "true"
     include("protocol/handlers.jl")
     include("protocol/parameters.jl")
     include("protocol/test_versioning.jl")
+    include("protocol/test_modern.jl")
     include("utils/serialization.jl")
     include("utils/logging.jl")
     include("transports/test_stdio.jl")